+++ float.sh
@@ -1,16 +1,15 @@
 #!/bin/bash
 
 # > float.sh <
-# Floating point number functions.
+# Support for floating point math in bash.
 
-# This code is republished from Linux Journal, July 30, 2008:
+# The original code appears on linuxjournal.com, July 30, 2008:
 # http://www.linuxjournal.com/content/floating-point-math-bash
 
-# It was written by Mitch Frazier (I think) and covered by the GNU Public License, version 2.
+# These were written by Mitch Frazier (I think) and are covered by the GNU Public License, version 2.
 
-#####################################################################
-# Default scale used by float functions.
 
+# Default scale used by bc.
 float_scale=2
 
 # Evaluate a floating point number expression.
@@ -20,7 +19,9 @@ function float_eval () {
   if [[ $# -gt 0 ]]; then
     result=$(echo "scale=$float_scale; $*" | bc -q 2>/dev/null)
     stat=$?
-    if [[ $stat -eq 0  &&  -z "$result" ]]; then stat=1; fi
+    if [[ $stat -eq 0  &&  -z "$result" ]]; then
+      stat=1;
+    fi
   fi
   echo $result
   return $stat
@@ -31,8 +32,12 @@ function float_cond () {
   local cond=0
   if [[ $# -gt 0 ]]; then
     cond=$(echo "$*" | bc -q 2>/dev/null)
-    if [[ -z "$cond" ]]; then cond=0; fi
-    if [[ "$cond" != 0  &&  "$cond" != 1 ]]; then cond=0; fi
+    if [[ -z "$cond" ]];
+      then cond=0;
+    fi
+    if [[ "$cond" != 0  &&  "$cond" != 1 ]]; then
+      cond=0;
+    fi
   fi
   local stat=$((cond == 0))
   return $stat